# back/userAPI/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Back-end do Projeto Pilha Completa",
    "version": "1.0.0",
    "title": "Projeto Pilha Completa"
  },
  "schemes": [
    "https"
  ],
  "paths": {
    "/getusers": {
      "get": {
        "summary": "Obtem todos os usuários",
        "operationId": "getUsers",
        "description": "Retorna todos os usuários cadastrados",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Retorna todos os Usuários",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "500": {
            "description": "Erro interno."
          }
        }
      }
    },
    "/getuser/:id": {
      "get": {
        "summary": "Obtem um usuário específico",
        "operationId": "getUser",
        "description": "Obtem um usuário a partir do seu id",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado e retornado",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/createuser": {
      "post": {
        "summary": "Cria um usuário",
        "operationId": "createUser",
        "description": "Cria um usuário com todos os dados requeridos",
        "tags": [
          "User"
        ],
        "responses": {
          "201": {
            "description": "Criado",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "500": {
            "description": "Erro interno."
          }
        }
      }
    },
    "/updateuser/:id": {
      "put": {
        "summary": "Atualiza um usuário especifico",
        "operationId": "updateUser",
        "description": "Atualiza um usuário específico identificado por seu id",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Atualizado"
          },
          "500": {
            "description": "Acesso negado"
          }
        }
      }
    },
    "/deleteuser/:id": {
      "delete": {
        "summary": "Apaga um usuário",
        "operationId": "deleteUser",
        "description": "Apaga um usuário identificado por seu id",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Deletado"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "dadosPessoais": {
          "type": "object",
          "properties": {
            "nome": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "login": {
              "type": "string"
            }
          },
          "required": [
            "nome",
            "email",
            "login"
          ]
        }
      }
    }
  }
}